Fix circular dependency for Parallel/PhaseControl
