+++ src/Evolution/Executables/GeneralizedHarmonic/CMakeLists.txt
@@ -31,6 +31,7 @@ set(LIBS_TO_LINK
   MathFunctions
   Options
   Parallel
+  PhaseControl
   Time
   Utilities
   )
