+++ src/Evolution/Executables/ScalarWave/CMakeLists.txt
@@ -16,6 +16,7 @@ set(LIBS_TO_LINK
   Time
   Options
   Parallel
+  PhaseControl
   Utilities
   WaveEquationSolutions
   )
