+++ src/Parallel/PhaseControl/CMakeLists.txt
@@ -1,6 +1,10 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
+set(LIBRARY PhaseControl)
+
+add_spectre_library(${LIBRARY} INTERFACE)
+
 spectre_target_headers(
   ${LIBRARY}
   INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
@@ -9,4 +13,14 @@ spectre_target_headers(
   PhaseChange.hpp
   PhaseControlTags.hpp
   VisitAndReturn.hpp
+  )
+
+target_link_libraries(
+  ${LIBRARY}
+  INTERFACE
+  DataStructures
+  EventsAndTriggers
+  Options
+  Parallel
+  Utilities
   )
